backend/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
beautifulsoup4
wikipedia
graphviz
sqlalchemy
pgvector
psycopg[binary]
pypdf
pandas
numpy
ollama
typing
matplotlib
networkx
streamlit
google-generativeai
youtube-transcript-api
pathlib
langchain_groq
flask
diffusers 
transformers 
torch